% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_map.R
\name{distance_map}
\alias{distance_map}
\title{calculate the distance to the nearest 5' or 3' subject GRanges}
\usage{
distance_map(query_gr, subject_gr, end = c("five", "three"),
  maximum_distance = 2000, standardize = T)
}
\arguments{
\item{query_gr}{The query GRanges.}

\item{subject_gr}{The subject GRanges.}

\item{end}{Can be "five" or "three", indicating the distance to the five prime end or three prime end.}

\item{maximum_distance}{The maximum distance returned, the un-defined distances are set to this value also, default is 2000.}

\item{standardize}{wheather standardize the output distance, i.e. minus the means and dividing the standard deviations.}
}
\value{
a vector of mapped distances with the length equal to \code{query_gr}.
}
\description{
\code{distance_map} calculate the distance to the nearest 5' or 3' end.
}
