% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_longest_transcript.R
\name{find_longest_transcript}
\alias{find_longest_transcript}
\title{Find the longest transcript given an GRangesList object obtained by exonsByTranscripts.}
\usage{
find_longest_transcript(exbtx, txdb)
}
\arguments{
\item{exbtx}{a GRangesList object obtaied from the exonsByTranscripts function defined in package GenomicFeatures.}

\item{txdb}{the corresponding txdb object.}
}
\value{
a data.frame indicating a given TXID is longest for the gene or not.
}
\description{
\code{find_longest_transcript} is used to find the longest transcript per gene.
}
