% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_regular.R
\name{glm_regular}
\alias{glm_regular}
\title{Build and visualize a generalized linear model.}
\usage{
glm_regular(Y, PREDICTORS, HDER = "glm", family = c("gaussian",
  "binomial", "poisson"), CUT_OFF = 5, Critical_value = 0.05,
  Exclude_intercept = F, Sort_by = c("byZstat", "byLogit"))
}
\arguments{
\item{Y}{A \code{vector} that defines the response variable. It can be a 2-column integer matrix given the binomial family：the first column gives the number of successes and the second the number of failures.}

\item{PREDICTORS}{A \code{data.frame} that defines the model design, which will include the predictors / features in the collumns.}

\item{HDER}{The subtitle and the file name of the plot.}

\item{family}{Define the family of the glm, should be one of "gaussian", "binomial", and "poisson",}

\item{CUT_OFF}{The cut off of the occurence of the less abundence class in binary features, if the less frequent class is less than this threshold, the feature will be dropped, default is 5. This is important when we want to have a reliable asymptotics result in Wald test.}

\item{Critical_value}{The critical value used on adjusted p values, default is 0.05.}

\item{Exclude_intercept}{Whether to omit the intercept term when plot the estimates and statistics, this should be applied when the intercept estimates is too big relative to other predictors, default is FALSE.}

\item{Sort_by}{Determine the order of the predictors showed in the plot, should be one of "byLogit", "byZstat", default is byLogit.}
}
\description{
\code{glm_regular} is a function used to build a regular generalized linear model based on genomic features.
}
\details{
The function will fit a linear model based on the provided predictors and the response variable. The coefficient estimates and the wald test statistics will be saved in a graph.
}
