% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_multinomial.R
\name{go_multinomial}
\alias{go_multinomial}
\title{Run GO analysis over methylation sets with multi-class labeling.}
\usage{
go_multinomial(Y, row_ranges, txdb, orgDb, HDER = "GO_multinomial",
  category = "BP", gene_key = "ENTREZID", min_bg_count = 10,
  min_gs_count = 10, max_bg_count = 500, max_gs_count = 500,
  pvalue_correction = "BH", GO_Slim = FALSE, EASE_Score = TRUE,
  top_terms = 10, save_chart = FALSE, background_ranges = NULL)
}
\arguments{
\item{Y}{A \code{vector} that defines the observed multinomial response variable.}

\item{row_ranges}{A \code{GRanges} object that defines the genomic location of each sites on genome.}

\item{txdb}{A \code{txdb} object downloaded from bioconductor.}

\item{orgDb}{An \code{OrgDb} object defined by AnnotationDbi.}

\item{HDER}{A \code{character} defining the subtitle and the file name of the plot.}

\item{category}{A character specifying the gene ontology category, can be one of "BP", "CC", and "MF", default "BP".}

\item{gene_key}{A character specifying the type of the gene ID, the available types of the keys can be find using \code{keytypes(org.Hs.eg.db)}, default "ENTREZID".}

\item{min_bg_count}{Term minimum number of occurence in background genes; default 10.}

\item{min_gs_count}{Term minimum number of occurence in gene set genes; default 10.}

\item{max_bg_count}{Term maximum number of occurence in background genes; default 500.}

\item{max_gs_count}{Term maximum number of occurence in gene set genes; default 500.}

\item{pvalue_correction}{Method used for multiple hypothesis adjustment, can be one in "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", and "none".}

\item{GO_Slim}{Whether to conduct enrichment analysis only on GO slim terms (a certain subset to GO terms), default FALSE.}

\item{EASE_Score}{Whether or not use EASE score method. (a more conservative hypergeomatric test calculation used in DAVID) for more details please refer to \href{https://david.ncifcrf.gov/helps/functional_annotation.html#fisher}{here}, default TRUE.}

\item{top_terms}{Top number of terms (sorted by p values) for each multinomial class returned in the analysis, default is 10.}

\item{save_chart}{Whether to save the chart of GO enrichment for each calss, default FALSE.}

\item{background_ranges}{A \code{GRanges} object that defines the GO enrichment background, default is not used.}
}
\description{
\code{go_multinomial} is a function used to run GO analysis on multi-class labeled ranged data. The back ground used is the shared genes in all of the ranges.
}
