% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_loop.R
\name{infer_loop}
\alias{infer_loop}
\title{Infer looped RNA 2ndary structures given provided hybridized 2ndary structures.}
\usage{
infer_loop(struc_sterm, length_cut_off = 30)
}
\arguments{
\item{struc_sterm}{a \code{GRanges} or \code{GRangesList} object that indicates the hybridized structure region on each transcript.}

\item{length_cut_off}{the allowed maximum pairing distance between 2 hybridized regions, the default setting is 30.
Due to the crude algorithm defined by this function, The higher this value is, the more false positive loops are presented in the result.}
}
\value{
a \code{GRanges} object with regions corresponding to the inferred looped region given the provided hybridized region.
}
\description{
\code{infer_loop} is used to extract the roughly accurate looped regions beween closed hybridized pairs given regions of hybridization.
}
