% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.modeling.R
\name{logistic.modeling}
\alias{logistic.modeling}
\title{Conduct model selection, inference, and visualization for logistic regression.}
\usage{
logistic.modeling(se, beta_prior = "robust", model_prior = beta.binomial(1,
  1), MCMC_iterations = 1e+05, decision_method = "BPM", top = NULL,
  Model_set_name = "LogisticModel",
  Sample_Names_coldata = colnames(colData(se))[1])
}
\arguments{
\item{se}{A \code{\link{SummarizedExperiment}} object containing the matrix of response variables, each row should represent one modification site, and each collumn should represent a sample or condition. The entries of \code{assay} are integer values of 1 or 0, with 1 indicating the positive class and 0 indicating the negative class used in logistic regression.

The sample names should be defined by the \code{colnames} of the \code{SummarizedExperiment}, alternatively, it can be defined by the first collumn of the \code{colData}, or by the user defined collumns in \code{colData} by the argument \code{Sample_Names_coldata}.

The features or design matrix should be included in the \code{mcols} of the \code{SummarizedExperiment} object.}

\item{beta_prior}{prior distribution for regression coefficients, default setting is "robust"; see \code{\link{bas.glm}}}

\item{model_prior}{Family of prior distribution on the models, default setting is beta.binomial(1,1); see \code{\link{bas.glm}}}

\item{MCMC_iterations}{Number of models to sample, default setting is 100000,the maximum number is 1^e8.}

\item{decision_method}{Decision method used in the bayesian model selection, default setting is 'BPM'; see \code{\link{predict.bas}}}

\item{top}{The number of top models used for BMA related decision making, default setting is NULL (BMA over all models); see \code{\link{predict.bas}}}

\item{Model_set_name}{The name of the directory to save the statistics of the logistic models and their diagrams, the default name is "LogisticModel".}

\item{Sample_Names_coldata}{Provided column names in \code{colData} for the sample labels.}
}
\value{
A folder under the current working directory will be created, the reports and the diagrams will be saved within it.
}
\description{
\code{logistic.modeling} is used to perform logistic regression analysis based on previously generated features, it reports logistic regression statistics after model selection,
and it can plot diagrams of logit estimates, wald test / likelihood ratio test statistics, and the goodness of fit (deviance) across various samples.
}
\details{
\code{logistic.modeling} build logistic regression model on the features and targets defined by the annotated \code{\link{SummarizedExperiment}} object returned by \code{\link{predictors.annot}}. The model selection is conducted by bayesian model selection method defined by the package \code{\link{BAS}}.
}
\seealso{
Use \code{\link{predictors.annot}} to annotate features.
}
