% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties_map.R
\name{properties_map}
\alias{properties_map}
\title{Dealing with ambiguities of the GRanges in query that may mapped to multiple features with different properties.}
\usage{
properties_map(query_gr, feature_gr, feature_properties, no_map_val = NA,
  normalize = F)
}
\arguments{
\item{query_gr}{The GRanges to query.}

\item{feature_properties}{The vector of properties of the subject (e.x. the length of the Gene).}

\item{no_map_val}{The value give to the not mapped query, by default is NA.}

\item{normalize}{wheather standardize the output mapped properties, i.e. minus the means and dividing the SDs.}

\item{feature_grl}{The GRangesList to map on, e.x. the Genes derived by exonsby(txdb ,by = "genes").}
}
\value{
a vector of mapped properties with the length equal to \code{query_gr}.
}
\description{
\code{properties_map} mitigates the ambiguities by averaging the properties of multiple mapped query.
}
