% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relative_Pos_map.R
\name{Relative_Pos_map}
\alias{Relative_Pos_map}
\title{Calculate the relative position of a GRanges object on transcript.}
\usage{
Relative_Pos_map(query_gr, feature_grl, no_map_val = NA)
}
\arguments{
\item{query_gr}{The GRanges to query.}

\item{feature_grl}{The GRangesList to map on.}

\item{no_map_val}{The value give to the not mapped query, by default is NA.}
}
\value{
a numeric vector with value between [0,1].
}
\description{
\code{Relative_Pos_map} is used to calculate the relative position of a GRanges object on transcript regions defined by a GRangesList object.
The query GRanges is assumed to have a single based resolution (or width = 1), otherwise only the start location is considered.
}
