% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sequence.R
\name{sample_sequence}
\alias{sample_sequence}
\title{Sample nucleotides or motifs on a subset of genome defined by \code{GRanges}.}
\usage{
sample_sequence(Sequence, Subset_GR, BSgnm, Fixed = F, N = NULL,
  Replace = F)
}
\arguments{
\item{Sequence}{a character string indicating the query sequence or motifs; Rules in \code{\link{IUPAC_CODE_MAP}} is supported when \code{Fixed} = FALSE.}

\item{Subset_GR}{the \code{\link{GRanges}} object used to define the subset of the genome.}

\item{BSgnm}{the \code{\link{BSgenome}} object containing the sequence of the genome.}

\item{Fixed}{FALSE to support the vague mapping rule of the chagracter string, default is FALSE.}

\item{N}{number of ranges sampled, by default, it returns all the ranges without sampling.}

\item{Replace}{whether sample with replacement or not, default is FALSE.}
}
\value{
a \code{GRanges} object contains the (sampled) mapped regions of the query sequence on the defined subset of the genome.
}
\description{
\code{sample_sequence} is used to extract the regions of a query sequence mapped to user defined sub-regions of Genome.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)
exons_hg19 <- exons(TxDb.Hsapiens.UCSC.hg19.knownGene)
Motif_exons_hg19 <- Sample_sequence_on_TXDB("RRACH",exons_hg19,Hsapiens,N = 100000)
#The command above sample 100000 regions of RRACH on exonic regions of hg19.

}
