% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Struc_mm10}
\alias{Struc_mm10}
\title{MEA RNA 2ndary structures of mm10 exons predicted by RNAfold}
\format{A GRangesList object of length 34597:
\describe{
  Each element of the GRangesList represents a transcript, exept that the first element refers to the mitochondria chromosome.
  The GRanges within each of the GRangesList element represent the genomic locations of the MEA RNA 2ndary structures predicted by the method above.
}}
\usage{
rBS_gr = GRanges(seqnames = rBS_df$`#SeqID`, strand = rBS_df$refStrand, ranges = IRanges(start = rBS_df$refPos, width = 1))
rBS_gr_filtered <- rBS_gr[!rBS_gr \%over\% Struc_mm10]
}
\description{
The full length transcripts are extracted from mm10;
the transcript annotation file was downloaded from the refSeq archive-2015-07-17-14-33-26.
The temperature of RNAfold prediction was set at 70 degree, and
the gammar for MEA structure was set at 0.1. The maximum paring distance was set at 150bp.
}
\details{
For the transcripts longer than 8000bp, the predictions were conducted in windows of 2000bp with steps of 1000bp.
}
\keyword{datasets}
